requirements.txt
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
asn1crypto==1.5.1
attrs==23.2.0
bcrypt==4.1.2
boto3==1.34.88
botocore==1.34.88
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coincurve==18.0.0
colorclass==2.2.2
cryptography==42.0.5
cssselect==1.2.0
cssutils==2.10.2
cytoolz==0.12.3
distlib==0.3.8
dnspython==2.6.1
docopt==0.6.2
email_validator==2.1.1
eth-hash==0.7.0
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-typing==3.5.2
eth-utils==2.3.1
exceptiongroup==1.2.1
fastapi==0.110.2
filelock==3.13.4
frozenlist==1.4.1
grpcio==1.62.2
gunicorn==22.0.0
h11==0.14.0
httptools==0.6.1
iconsdk==2.5.1
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.2.1
msgpack==1.0.8
multimethod==1.9.1
multipledispatch==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
packaging==24.0
pip-upgrader==1.4.15
platformdirs==4.2.0
pluggy==1.4.0
premailer==3.10.0
protobuf==5.26.1
py==1.11.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.0
pydantic_core==2.18.1
PyJWT==2.8.0
PyMySQL==1.1.0
pyparsing==3.1.2
pyrsistent==0.20.0
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
ray==2.10.0
referencing==0.34.0
requests==2.28.2
requests-mock==1.10.0
rpds-py==0.18.0
s3transfer==0.10.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.37.2
termcolor==2.4.0
terminaltables==3.1.10
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
typing_extensions==4.11.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.25.3
watchgod==0.8.2
websocket-client==1.5.3
websockets==12.0
yagmail==0.15.293
zipp==3.18.1
